refactor(gui): move quick download logic to a separate component
